[09] Fix third Devin Review round
- H.264/H.265 AVCC/HVCC NAL length overflow checks now use u64 shifts,
  avoiding panic/wrap on 32-bit wasm32 targets with 4-byte length size.
- MP3 frame length now computes slots first and multiplies by slot_size,
  giving the correct ISO 11172-3 truncation for Layer I (e.g. 136 bytes
  for 128kbps / 44.1kHz instead of 139).
- Add padding case to the Layer I regression test and update expected value.

+++ crates/cheetah-media-bitstream/src/h264.rs
@@ -87,7 +87,7 @@ impl<'a> NalUnit<'a> {
             return Err(H264Error::InvalidNalLength);
         }
         let len = self.data.len();
-        if len >= (1usize << (length_size * 8)) {
+        if (len as u64) >= (1u64 << (length_size * 8)) {
             return Err(H264Error::InvalidNalLength);
         }
         let mut out = Vec::with_capacity(usize::from(length_size) + self.data.len());
